+++ include/iutest_ver.hpp
@@ -17,11 +17,11 @@
 
 //======================================================================
 // define
-#define IUTEST_VER          0x01169912u //!< iutest version 1.16.99.12
+#define IUTEST_VER          0x01169913u //!< iutest version 1.16.99.13
 #define IUTEST_MAJORVER     0x01u       //!< Major Version
 #define IUTEST_MINORVER     0x16u       //!< Minor Version
 #define IUTEST_MICROVER     0x99u       //!< Micro Version
-#define IUTEST_REVISION     0x12u       //!< Revision
+#define IUTEST_REVISION     0x13u       //!< Revision
 
 #define IUTEST_BUILD        IUTEST_MICROVER //!< @deprecated
 
@@ -123,6 +123,7 @@ modification, are permitted provided that the following conditions are met: \n
   <li>v1.17.0.0
     <ul>
       <li></li>
+      <li>Visual Studio 2019 対応</li>
       <li>junit xml 出力で出力するテストが 0 だった場合にファイル書き出ししないように修正</li>
       <li>CSV パラメータ生成でファイルオープン失敗した場合の警告レベルを FATAL から WARNING に変更/li>
       <li>CSV パラメータ生成で要素がなかった場合に WARNING 出力</li>